% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvt-package.R
\docType{package}
\name{rvt-package}
\alias{rvt}
\alias{rvt-package}
\title{rvt: Relief Visualization Toolbox for R}
\description{
Pure R implementation of the Relief Visualization Toolbox (RVT) algorithms for terrain visualization from LiDAR data. Computes hillshade, slope, sky-view factor, positive openness, simple local relief model, and the Visualization for Archaeological Topography (VAT) composite blend. Faithful port of rvt-py (Kokalj et al.) with no Python dependency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nick-gauthier/rvt}
  \item Report bugs at \url{https://github.com/nick-gauthier/rvt/issues}
}

}
\author{
\strong{Maintainer}: Nick Gauthier \email{ngauthier@ufl.edu} (\href{https://orcid.org/0000-0001-6605-3470}{ORCID})

Authors:
\itemize{
  \item Anthropic Claude (R port of rvt-py algorithms)
}

}
\keyword{internal}
