% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{rvt_svf_compute}
\alias{rvt_svf_compute}
\title{Core SVF/Openness computation (internal, optimised)}
\usage{
rvt_svf_compute(
  height,
  r_max,
  r_min,
  n_dir,
  compute_svf = TRUE,
  compute_opns = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{height}{Padded height matrix (elevation * ve / resolution).}

\item{r_max}{Maximum search radius in pixels.}

\item{r_min}{Minimum search radius in pixels.}

\item{n_dir}{Number of search directions.}

\item{compute_svf}{Compute sky-view factor?}

\item{compute_opns}{Compute positive openness?}

\item{verbose}{Print progress?}
}
\value{
List with \verb{$svf} and/or \verb{$opns} matrices.
}
\description{
Core SVF/Openness computation (internal, optimised)
}
\keyword{internal}
