% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra.R
\name{slrm}
\alias{slrm}
\title{Simple Local Relief Model from a SpatRaster DEM}
\usage{
slrm(dem, radius_cell = 20L, ve_factor = 1)
}
\arguments{
\item{dem}{A \link[terra:SpatRaster-class]{terra::SpatRaster} (single layer).}

\item{radius_cell}{Mean filter radius in pixels (default 20).}

\item{ve_factor}{Vertical exaggeration (default 1).}
}
\value{
A \link[terra:SpatRaster-class]{terra::SpatRaster}.
}
\description{
Simple Local Relief Model from a SpatRaster DEM
}
